Ef Code First Config ist lecker

+++ Hallo_EFCodeFirst/Hallo_EFCodeFirst/Model/Tier.cs
@@ -5,7 +5,7 @@ namespace Hallo_EFCodeFirst.Model
     public class Tier : Entity
     {
         public string Art { get; set; }
-        public int AnzahlBeine { get; set; }
+        public byte AnzahlBeine { get; set; }
         public Tieroberfläche Oberfläche { get; set; }
         public HashSet<Mahlzeit> Mahlzeiten { get; set; } = new HashSet<Mahlzeit>();
